% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeders_equation.R
\name{breeders_equation}
\alias{breeders_equation}
\title{Perform the basic Breeder's Equation}
\usage{
breeders_equation(selection.differential, heritability)
}
\arguments{
\item{selection.differential}{= The change in the value of a polygenic trait within a generation.}

\item{heritability}{= The heritability of a polygenic trait.}
}
\description{
Perform the basic Breeder's Equation
}
