% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeders_equation_male_female.R
\name{breeders_equation_male_female}
\alias{breeders_equation_male_female}
\title{Perform the Breeder's Equation with separate selection differentials for males and females.}
\usage{
breeders_equation_male_female(
  heritability,
  female.selection.differential,
  male.selection.differential
)
}
\arguments{
\item{heritability}{= The heritability of a polygenic trait.}

\item{female.selection.differential}{= The selection differential for females}

\item{male.selection.differential}{= The selection differential for males}
}
\description{
Perform the Breeder's Equation with separate selection differentials for males and females.
}
