% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeders_equation_male_female_fitness.R
\name{breeders_equation_male_female_fitness}
\alias{breeders_equation_male_female_fitness}
\title{Calculate the response to selection from fitness costs only.}
\usage{
breeders_equation_male_female_fitness(
  male.fitness.selection.differential,
  female.fitness.selection.differential,
  heritability
)
}
\arguments{
\item{male.fitness.selection.differential}{= The selection differential for male mosquitoes from fitness costs}

\item{female.fitness.selection.differential}{= The selection differential for female mosquitoes from fitness costs}

\item{heritability}{= The heritability of a polygenic trait.}
}
\description{
Calculates the response to selection from fitness costs only, such as would occur when the
insecticide is either not deployed in the intervention site or the selection that occurs
in the refugia.
}
