% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/breeders_equation_male_female_insecticide_fitness.R
\name{breeders_equation_male_female_insecticide_fitness}
\alias{breeders_equation_male_female_insecticide_fitness}
\title{Calculate the response to selection from insecticide selection and fitness costs from males and females.}
\usage{
breeders_equation_male_female_insecticide_fitness(
  heritability,
  male.insecticide.fitness.selection.differential,
  female.insecticide.fitness.selection.differential
)
}
\arguments{
\item{heritability}{= The heritability of a polygenic trait.}

\item{male.insecticide.fitness.selection.differential}{= The selection differential for male mosquitoes following insecticide selection and fitness costs.}

\item{female.insecticide.fitness.selection.differential}{= The selection differential for female mosquitoes following insecticide selection and fitness costs.}
}
\description{
Calculate the response to selection from insecticide selection and fitness costs from males and females.
}
