% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_current_insecticide_concentration.R
\name{calculate_current_insecticide_concentration}
\alias{calculate_current_insecticide_concentration}
\title{Calculate the current insecticide concentration in deployment}
\usage{
calculate_current_insecticide_concentration(
  applied.concentration,
  instantaneous.decay.rate,
  generations.since.deployment
)
}
\arguments{
\item{applied.concentration}{= The concentration of the active ingredient that was initially deployed}

\item{instantaneous.decay.rate}{= The decay rate of the insecticide}

\item{generations.since.deployment}{= The number of mosquito generations since the insecticide was deployed.}
}
\description{
Calculate the current insecticide concentration in deployment
}
