% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_current_insecticide_efficacy.R
\name{calculate_current_insecticide_efficacy}
\alias{calculate_current_insecticide_efficacy}
\title{Calculate the current insecticidal killing efficacy against a fully susceptible mosquito.}
\usage{
calculate_current_insecticide_efficacy(
  generations.since.deployment,
  threshold.generations,
  initial.insecticide.efficacy,
  base.efficacy.decay.rate,
  rapid.decay.rate
)
}
\arguments{
\item{generations.since.deployment}{= the time in generations since the insecticide was last deployed.}

\item{threshold.generations}{= the time (in generations) after which there is a change in the insecticide effective dosage decay rate; which can be due to nets becoming worn to the point mosquitoes are able to avoid the insecticide.}

\item{initial.insecticide.efficacy}{= The efficacy of the applied insecticide concentration of the insecticide against a fully susceptible mosquito}

\item{base.efficacy.decay.rate}{= The efficacy decay rate during the first generations during which there is expected to be a slow decay in insecticidal efficacy}

\item{rapid.decay.rate}{= Decay rate of insecticidal efficacy decay after the threshold.generations is exceeded, where the insecticidal product is expected to undergo a rapid change in efficacy.}
}
\description{
Calculate the current insecticidal killing efficacy against a fully susceptible mosquito.
}
