% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_density_of_trait_values.R
\name{calculate_density_of_trait_values}
\alias{calculate_density_of_trait_values}
\title{Calculate the relative population contribution to the population of mosquitoes with a Polygenic Resistance Score prior to selection.}
\usage{
calculate_density_of_trait_values(
  vector.length,
  trait.mean,
  standard.deviation
)
}
\arguments{
\item{vector.length}{= The length of the vector to be returned. A minimum value of 100,000 is recommmended.}

\item{trait.mean}{= The mean value of a polygenic trait in a population.}

\item{standard.deviaton}{= The standard deviation of the trait mean in the population.}
}
\description{
Calculate the relative population contribution to the population of mosquitoes with a Polygenic Resistance Score prior to selection.
}
