% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_density_unexposed.R
\name{calculate_density_unexposed}
\alias{calculate_density_unexposed}
\title{Calculate the relative population contribution to the population of mosquitoes with a Polygenic Resistance Score that did not encounter the insecticide in the intervention site.}
\usage{
calculate_density_unexposed(
  vector.length,
  trait.mean,
  standard.deviation,
  insecticide.exposure
)
}
\arguments{
\item{vector.length}{= The length of the vector to be returned. A minimum value of 100,000 is recommmended.}

\item{trait.mean}{= The mean value of a polygenic trait in a population.}

\item{standard.deviation}{= The standard deviation of the trait mean in the population.}

\item{insecticide.exposure}{= The proportion of the mosquito population in the intervention site being exposed to an insecticide.}
}
\description{
Calculate the relative population contribution to the population of mosquitoes with a Polygenic Resistance Score that did not encounter the insecticide in the intervention site.
}
