% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_female_fitness_selection_differential.R
\name{calculate_female_fitness_selection_differential}
\alias{calculate_female_fitness_selection_differential}
\title{Calculate the selection differential for female mosquitoes from fitness costs}
\usage{
calculate_female_fitness_selection_differential(
  female.trait.mean,
  female.trait.mean.after.fitness
)
}
\arguments{
\item{female.trait.mean}{= The mean Polygenic Resistance Score to the insecticide for female mosquitoes before selection has occurred.}

\item{female.trait.mean.after.fitness}{= The Mean Female Polygenic Resistance Score of the population after fitness costs have been applied.}
}
\description{
Calculate the selection differential for female mosquitoes from fitness costs
}
