% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_female_insecticide_fitness_selection_differential.R
\name{calculate_female_insecticide_fitness_selection_differential}
\alias{calculate_female_insecticide_fitness_selection_differential}
\title{Calculate the selection differential for female mosquitoes following insecticide selection and fitness costs.}
\usage{
calculate_female_insecticide_fitness_selection_differential(
  female.insecticide.selection.differential,
  exposure.scaling.factor,
  female.fitness.selection.differential
)
}
\arguments{
\item{female.insecticide.selection.differential}{= The selection differential for female mosquitoes following insecticide selection.}

\item{exposure.scaling.factor}{= A factor which converts the insecticide exposure to the selection differential.}

\item{female.fitness.selection.differential}{= The selection differential for female mosquitoes from fitness costs}
}
\description{
Calculate the selection differential for female mosquitoes following insecticide selection and fitness costs.
}
