% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_female_insecticide_selection_differential.R
\name{calculate_female_insecticide_selection_differential}
\alias{calculate_female_insecticide_selection_differential}
\title{Calculate the selection differential for female mosquitoes following insecticide selection.}
\usage{
calculate_female_insecticide_selection_differential(
  female.trait.mean.after.selection,
  female.trait.mean
)
}
\arguments{
\item{female.trait.mean.after.selection}{= The mean Polygenic Resistance Score to the insecticide of the female mosquitoes after insecticide selection.}

\item{female.trait.mean}{= The mean Polygenic Resistance Score to the insecticide of female mosquitoes before selection has occurred.}
}
\description{
Calculate the selection differential for female mosquitoes following insecticide selection.
}
