% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_female_population_size_after_selection.R
\name{calculate_female_population_size_after_selection}
\alias{calculate_female_population_size_after_selection}
\title{Calculate the relative population size of all the female mosquitoes in the intervention site after insecticide selection.}
\usage{
calculate_female_population_size_after_selection(
  female.population.size.unexposed,
  female.population.size.exposed.survivors
)
}
\arguments{
\item{female.population.size.unexposed}{= The relative population size of the unexposed female mosquitoes in the intervention site.}

\item{female.population.size.exposed.survivors}{= The relative population size of female mosquitoes that have encountered the insecticide and survived.}
}
\description{
Calculate the relative population size of all the female mosquitoes in the intervention site after insecticide selection.
}
