% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_female_population_size_unexposed.R
\name{calculate_female_population_size_unexposed}
\alias{calculate_female_population_size_unexposed}
\title{Calculate the relative population size of the unexposed female mosquitoes in the intervention site.}
\usage{
calculate_female_population_size_unexposed(
  total.female.population.size,
  female.insecticide.exposure
)
}
\arguments{
\item{total.female.population.size}{= The total relative population size of female mosquitoes in the intervention site before insecticide selection.}

\item{female.insecticide.exposure}{= Proportion of female mosquitoes in the intervention site that encounter and are exposed to the deployed insecticide.}
}
\description{
Calculate the relative population size of the unexposed female mosquitoes in the intervention site.
}
