% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_female_trait_mean_after_fitness.R
\name{calculate_female_trait_mean_after_fitness}
\alias{calculate_female_trait_mean_after_fitness}
\title{Calculate The Mean Female Polygenic Resistance Score of the population after fitness costs have been applied.}
\usage{
calculate_female_trait_mean_after_fitness(
  female.trait.mean,
  female.fitness.cost
)
}
\arguments{
\item{female.trait.mean}{= The mean Polygenic Resistance Score to the insecticide for female mosquitoes before selection has occurred.}

\item{female.fitness.cost}{= The fixed fitness cost associated with polygenic resistance for females.}
}
\description{
Calculate The Mean Female Polygenic Resistance Score of the population after fitness costs have been applied.
}
