% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_female_trait_mean_post_selection.R
\name{calculate_female_trait_mean_post_selection}
\alias{calculate_female_trait_mean_post_selection}
\title{Calculate the mean Polygenic Resistance Score to the insecticide for female mosquitoes after insecticide selection.}
\usage{
calculate_female_trait_mean_post_selection(
  female.population.size.exposed.survivors,
  female.trait.mean.exposed.survivors,
  female.population.size.unexposed,
  female.trait.mean,
  female.population.size.after.selection
)
}
\arguments{
\item{female.population.size.exposed.survivors}{= The relative population size of female mosquitoes that have encountered the insecticide and survived.}

\item{female.trait.mean.exposed.survivors}{= The mean Polygenic Resistance Score of the female mosquitoes exposed to and surviving the insecticide encounter.}

\item{female.population.size.unexposed}{= The relative population size of the unexposed female mosquitoes in the intervention site.}

\item{female.trait.mean}{= The mean Polygenic Resistance Score to the insecticide of female mosquitoes before selection has occurred.}

\item{female.population.size.after.selection}{= The relative population size of all the female mosquitoes in the intervention site after insecticide selection.}
}
\description{
Calculate the mean Polygenic Resistance Score to the insecticide for female mosquitoes after insecticide selection.
}
