% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fitness_selection_differential.R
\name{calculate_fitness_selection_differential}
\alias{calculate_fitness_selection_differential}
\title{Calculate the selection differential as a result of fitness costs.}
\usage{
calculate_fitness_selection_differential(trait.mean.after.fitness, trait.mean)
}
\arguments{
\item{trait.mean.after.fitness}{= The mean Polygenic Resistance Score after fitness costs are applied to the population.}

\item{trait.mean}{= The mean value of a polygenic trait in a population.}
}
\description{
Calculate the selection differential as a result of fitness costs.
}
