% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_fitness_selection_differential_male_female.R
\name{calculate_fitness_selection_differential_male_female}
\alias{calculate_fitness_selection_differential_male_female}
\title{Calculate the fitness selection differential from separate male and female selection differential inputs}
\usage{
calculate_fitness_selection_differential_male_female(
  male.fitness.selection.differential,
  female.fitness.selection.differential
)
}
\arguments{
\item{male.fitness.selection.differential}{=The selection differential for males following from fitness costs.}

\item{female.fitness.selection.differential}{= The selection differential for female mosquitoes from fitness costs}
}
\description{
Calculate the fitness selection differential from separate male and female selection differential inputs
}
