% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_half_population_survival.R
\name{calculate_half_population_survival}
\alias{calculate_half_population_survival}
\title{Calculate the Half Population Survival for a given Polygenic Resistance Score scale.}
\usage{
calculate_half_population_survival(
  desired.polygenic.resistance.score = 100,
  desired.survival.proportion = 0.1,
  maximum.bioassay.survival.proportion = 1,
  michaelis.menten.slope = 1,
  estimate.precision = 0.001,
  minimum.resistance.value = 0,
  maximum.resistance.value = 25000
)
}
\arguments{
\item{desired.polygenic.resistance.score}{The value you want a survival value to correspond to}

\item{desired.survival.proportion}{The survival proportion you want your desired.polygenic.resistance.score to have. Values must be between 0 and 1}

\item{maximum.bioassay.survival.proportion}{Must be set at 1}

\item{michaelis.menten.slope}{Must be set at 1}

\item{estimate.precision}{How precise your estimate of insecticide restistance intensity should be. Recommend values between 0.01 to 0.001}

\item{minimum.resistance.value}{Recommend setting to 0. Must be lower than half survival resistance.}

\item{maximum.resistance.value}{Depends on your scale. Recommend setting arbitrarily high (10000). Must be higher than half resistance survival}
}
\value{
a value for the half.population.survival.value which is the half.population.bioassay.survival.resistance for the scale.
}
\description{
A function that enables calculation of the  50\% survival threshold, based on desired
resistance value and its desired survival value. For example if you want to base the scale on
a 20\% survival having a Polygenic Resistance Score of 1000.
}
