% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_initial_efficacy_from_insecticide_concentration.R
\name{calculate_initial_efficacy_from_insecticide_concentration}
\alias{calculate_initial_efficacy_from_insecticide_concentration}
\title{Calculate the initial insecticidal efficacy from the insecticide concentrations}
\usage{
calculate_initial_efficacy_from_insecticide_concentration(
  applied.insecticide.dose,
  recommended.insecticide.dose
)
}
\arguments{
\item{applied.insecticide.dose}{= The actual applied dosage of the insecticide deployed.}

\item{recommended.insecticide.dose}{= The manufacturer’s recommended dosage for the insecticide when deployed as a single insecticide formulation.}
}
\description{
Calculate the initial insecticidal efficacy from the insecticide concentrations
}
