% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_initial_efficacy_from_insecticide_concentration_mixtures.R
\name{calculate_initial_efficacy_from_insecticide_concentration_mixtures}
\alias{calculate_initial_efficacy_from_insecticide_concentration_mixtures}
\title{Calculate the initial insecticide efficacy when the insecticide is deployed in a mixture formulation}
\usage{
calculate_initial_efficacy_from_insecticide_concentration_mixtures(
  mixture.dose,
  recommended.insecticide.dose
)
}
\arguments{
\item{mixture.dose}{= The deployed insecticide dosage when the insecticide is deployed in a mixture formulation.}

\item{recommended.insecticide.dose}{= The manufacturer’s recommended dosage for the insecticide when deployed as a single insecticide formulation.}
}
\description{
Calculate the initial insecticide efficacy when the insecticide is deployed in a mixture formulation
}
