% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_insecticide_fitness_selection_differential.R
\name{calculate_insecticide_fitness_selection_differential}
\alias{calculate_insecticide_fitness_selection_differential}
\title{Calculate the overall selection differential as a result of both insecticide selection and fitness costs}
\usage{
calculate_insecticide_fitness_selection_differential(
  exposure.scaling.factor,
  insecticide.selection.differential,
  fitness.selection.differential
)
}
\arguments{
\item{exposure.scaling.factor}{= A factor which converts the insecticide exposure to the selection differential.}

\item{insecticide.selection.differential}{= The selection differential as a result of insecticide selection.}

\item{fitness.selection.differential}{= The selection differential as a result of fitness costs.}
}
\description{
Calculate the overall selection differential as a result of both insecticide selection and fitness costs
}
