% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_insecticide_population_suppression.R
\name{calculate_insecticide_population_suppression}
\alias{calculate_insecticide_population_suppression}
\title{The impact of insecticides on the relative population size of female mosquitoes the intervention site, who are now ready to lay eggs.}
\usage{
calculate_insecticide_population_suppression(
  female.population.size.after.selection,
  total.female.population.size
)
}
\arguments{
\item{female.population.size.after.selection}{= The relative population size of all the female mosquitoes in the intervention site after insecticide selection.}

\item{total.female.population.size}{= The total relative population size of female mosquitoes in the intervention site before insecticide selection.}
}
\description{
The impact of insecticides on the relative population size of female mosquitoes the intervention site, who are now ready to lay eggs.
}
