% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_insecticide_selection_differential.R
\name{calculate_insecticide_selection_differential}
\alias{calculate_insecticide_selection_differential}
\title{Calculate the selection differential as a result of insecticide selection.}
\usage{
calculate_insecticide_selection_differential(mean.post.selection, trait.mean)
}
\arguments{
\item{mean.post.selection}{= The mean Polygenic Resistance Score of the all the mosquitoes after insecticide selection has occurred.}

\item{trait.mean}{= The mean value of a polygenic trait.}
}
\description{
Calculate the selection differential as a result of insecticide selection.
}
