% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_intervention_site_after_selection_deployed.R
\name{calculate_intervention_site_after_selection_deployed}
\alias{calculate_intervention_site_after_selection_deployed}
\title{Calculate the mean Polygenic Resistance Score of the mosquito population in the intervention site after insecticide and fitness cost based selection.}
\usage{
calculate_intervention_site_after_selection_deployed(
  intervention.before.selection,
  response.insecticide.fitness
)
}
\arguments{
\item{intervention.before.selection}{= The mean Polygenic Resistance Score of the mosquito population in the intervention before selection has occurred that generation.}

\item{response.insecticide.fitness}{= The response to selection from insecticide selection and fitness costs.}
}
\description{
Calculate the mean Polygenic Resistance Score of the mosquito population in the intervention site after insecticide and fitness cost based selection.
}
