% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_intervention_site_after_selection_not_deployed.R
\name{calculate_intervention_site_after_selection_not_deployed}
\alias{calculate_intervention_site_after_selection_not_deployed}
\title{Calculate the mean Polygenic Resistance Score of the mosquito population in the intervention site after  fitness cost based selection.}
\usage{
calculate_intervention_site_after_selection_not_deployed(
  intervention.before.selection,
  response.fitness
)
}
\arguments{
\item{intervention.before.selection}{= The mean Polygenic Resistance Score of the mosquito population in the intervention before selection has occurred that generation.}

\item{response.fitness}{= The response to selection from fitness costs only, such as would occur when the insecticide is either not deployed in the intervention site or the selection that occurs in the refugia.}
}
\description{
Calculate the mean Polygenic Resistance Score of the mosquito population in the intervention site after  fitness cost based selection.
}
