% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_male_density_unexposed.R
\name{calculate_male_density_unexposed}
\alias{calculate_male_density_unexposed}
\title{Calculate the relative population contribution to the population of male mosquitoes with a Polygenic Resistance Score  who do not encounter the insecticide.}
\usage{
calculate_male_density_unexposed(
  vector.length,
  trait.mean,
  standard.deviation,
  female.insecticide.exposure,
  male.insecticide.exposure
)
}
\arguments{
\item{vector.length}{= The length of the vector to be returned. A minimum value of 100,000 is recommmended.}

\item{trait.mean}{= The mean value of a polygenic trait in a population.}

\item{standard.deviation}{= The standard deviation of the trait mean in the population.}

\item{female.insecticide.exposure}{= Proportion of female mosquitoes in the intervention site that encounter and are exposed to the deployed insecticide.}

\item{male.insecticide.exposure}{= Proportion of male mosquitoes in the intervention site that encounter and are exposed to the deployed insecticide as a proportion of the exposure of female mosquitoes}
}
\description{
Calculate the relative population contribution to the population of male mosquitoes with a Polygenic Resistance Score  who do not encounter the insecticide.
}
