% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_male_fitness_selection_differential.R
\name{calculate_male_fitness_selection_differential}
\alias{calculate_male_fitness_selection_differential}
\title{Calculate the selection differential for male mosquitoes from fitness costs}
\usage{
calculate_male_fitness_selection_differential(
  male.trait.mean,
  male.trait.mean.after.fitness
)
}
\arguments{
\item{male.trait.mean}{= The mean Polygenic Resistance Score to the insecticide for male mosquitoes before selection has occurred.}

\item{male.trait.mean.after.fitness}{= The mean male Polygenic Resistance Score of the population after fitness costs have been applied.}
}
\description{
Calculate the selection differential for male mosquitoes from fitness costs
}
