% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_male_insecticide_fitness_selection_differential.R
\name{calculate_male_insecticide_fitness_selection_differential}
\alias{calculate_male_insecticide_fitness_selection_differential}
\title{Calculate the selection differential for male mosquitoes following insecticide selection and fitness costs.}
\usage{
calculate_male_insecticide_fitness_selection_differential(
  male.insecticide.selection.differential,
  exposure.scaling.factor,
  male.fitness.selection.differential
)
}
\arguments{
\item{male.insecticide.selection.differential}{= The selection differential for male mosquitoes following insecticide selection.}

\item{exposure.scaling.factor}{= A factor which converts the insecticide exposure to the selection differential.}

\item{male.fitness.selection.differential}{= The selection differential for male mosquitoes from fitness costs}
}
\description{
Calculate the selection differential for male mosquitoes following insecticide selection and fitness costs.
}
