% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_male_insecticide_selection_differential.R
\name{calculate_male_insecticide_selection_differential}
\alias{calculate_male_insecticide_selection_differential}
\title{Calculate the selection differential for males following insecticide selection.}
\usage{
calculate_male_insecticide_selection_differential(
  male.trait.mean.after.selection,
  male.trait.mean
)
}
\arguments{
\item{male.trait.mean.after.selection}{= The mean Polygenic Resistance Score to the insecticide for male mosquitoes after insecticide selection.}

\item{male.trait.mean}{= The mean Polygenic Resistance Score to the insecticide for male mosquitoes before selection has occurred.}
}
\description{
Calculate the selection differential for males following insecticide selection.
}
