% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_male_population_size_after_selection.R
\name{calculate_male_population_size_after_selection}
\alias{calculate_male_population_size_after_selection}
\title{Calculate the relative population size of all the male mosquitoes in the intervention site after insecticide selection.}
\usage{
calculate_male_population_size_after_selection(
  male.population.size.exposed.survivors,
  male.population.size.unexposed
)
}
\arguments{
\item{male.population.size.exposed.survivors}{= The relative population size of male mosquitoes that have encountered the insecticide and survived.}

\item{male.population.size.unexposed}{= The relative population size of the unexposed male mosquitoes in the intervention site.}
}
\description{
Calculate the relative population size of all the male mosquitoes in the intervention site after insecticide selection.
}
