% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_male_population_size_exposed_survivors.R
\name{calculate_male_population_size_exposed_survivors}
\alias{calculate_male_population_size_exposed_survivors}
\title{Calculate the relative population size of male mosquitoes that have encountered the insecticide and survived.}
\usage{
calculate_male_population_size_exposed_survivors(
  relative.male.contributions.after.selection
)
}
\arguments{
\item{relative.male.contribution.after.selection}{= A vector of the relative population contributions to the population of male mosquitoes for all values of the Polygenic Resistance Score after the insecticide encounter in the intervention site.}
}
\description{
Calculate the relative population size of male mosquitoes that have encountered the insecticide and survived.
}
