% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_male_population_size_unexposed.R
\name{calculate_male_population_size_unexposed}
\alias{calculate_male_population_size_unexposed}
\title{Calculate the relative population size of the unexposed male mosquitoes in the intervention site.}
\usage{
calculate_male_population_size_unexposed(
  total.male.population.size,
  male.insecticide.exposure,
  female.insecticide.exposure
)
}
\arguments{
\item{total.male.population.size}{= The total relative population size of male mosquitoes in the intervention site before insecticide selection.}

\item{male.insecticide.exposure}{= Proportion of male mosquitoes in the intervention site that encounter and are exposed to the deployed insecticide as a proportion of the exposure of female mosquitoes}

\item{female.insecticide.exposure}{= Proportion of female mosquitoes in the intervention site that encounter and are exposed to the deployed insecticide.}
}
\description{
Calculate the relative population size of the unexposed male mosquitoes in the intervention site.
}
