% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_male_trait_mean_after_fitness.R
\name{calculate_male_trait_mean_after_fitness}
\alias{calculate_male_trait_mean_after_fitness}
\title{Calculate The Mean male Polygenic Resistance Score of the population after fitness costs have been applied.}
\usage{
calculate_male_trait_mean_after_fitness(male.trait.mean, male.fitness.cost)
}
\arguments{
\item{male.trait.mean}{= The mean Polygenic Resistance Score to the insecticide for male mosquitoes before selection has occurred.}

\item{male.fitness.cost}{= The fixed fitness cost associated with polygenic resistance for males.}
}
\description{
Calculate The Mean male Polygenic Resistance Score of the population after fitness costs have been applied.
}
