% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_male_trait_mean_post_selection.R
\name{calculate_male_trait_mean_post_selection}
\alias{calculate_male_trait_mean_post_selection}
\title{Calculate the mean Polygenic Resistance Score to the  insecticide for male mosquitoes after insecticide selection.}
\usage{
calculate_male_trait_mean_post_selection(
  male.population.size.exposed.survivors,
  male.trait.mean.exposed.survivors,
  male.population.size.unexposed,
  male.trait.mean,
  male.population.size.after.selection
)
}
\arguments{
\item{male.population.size.exposed.survivors}{= The relative population size of male mosquitoes that have encountered the insecticide and survived.}

\item{male.trait.mean.exposed.survivors}{= The mean Polygenic Resistance Score of the male mosquitoes exposed to and surviving the insecticide encounter.}

\item{male.population.size.unexposed}{= The relative population size of the unexposed male mosquitoes in the intervention site.}

\item{male.trait.mean}{= The mean Polygenic Resistance Score to the insecticide for male mosquitoes before selection has occurred.}

\item{male.population.size.after.selection}{= The relative population size of all the male mosquitoes in the intervention site after insecticide selection.}
}
\description{
Calculate the mean Polygenic Resistance Score to the  insecticide for male mosquitoes after insecticide selection.
}
