% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_population_size_post_selection.R
\name{calculate_population_size_post_selection}
\alias{calculate_population_size_post_selection}
\title{Calculate the relative population size of all the mosquitoes in the intervention site after selection has occurred.}
\usage{
calculate_population_size_post_selection(
  population.size.exposed.survivors,
  population.size.unexposed
)
}
\arguments{
\item{population.size.exposed.survivors}{= The relative population size of the mosquitoes in the intervention site who encountered the insecticide and survived.}

\item{population.size.unexposed}{= The relative population size of the mosquitoes in the intervention site who did not encounter the insecticide.}
}
\description{
Calculate the relative population size of all the mosquitoes in the intervention site after selection has occurred.
}
