% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_population_size_unexposed.R
\name{calculate_population_size_unexposed}
\alias{calculate_population_size_unexposed}
\title{Calculate the relative population size of the mosquitoes in the intervention site who did not encounter the insecticide.}
\usage{
calculate_population_size_unexposed(
  total.population.size,
  insecticide.exposure
)
}
\arguments{
\item{total.population.size}{= The relative population size of the intervention site prior to selection.}

\item{insecticide.exposure}{= The proportion of the mosquito population in the intervention site being exposed to an insecticide.}
}
\description{
Calculate the relative population size of the mosquitoes in the intervention site who did not encounter the insecticide.
}
