% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_selection_differential.R
\name{calculate_selection_differential}
\alias{calculate_selection_differential}
\title{Calculate the selection differential}
\usage{
calculate_selection_differential(trait.mean, trait.value.parents)
}
\arguments{
\item{trait.mean}{= The mean value of a polygenic trait before selection.}

\item{trait.value.parents}{= The mean polygenic trait value in the parents of the next generation.}
}
\description{
Calculate the selection differential
}
