% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_trait_mean_post_selection.R
\name{calculate_trait_mean_post_selection}
\alias{calculate_trait_mean_post_selection}
\title{Calculate the mean Polygenic Resistance Score of the all the mosquitoes after insecticide
selection has occurred.}
\usage{
calculate_trait_mean_post_selection(
  population.size.exposed.survivors,
  mean.score.exposed.survivors,
  population.size.unexposed,
  trait.mean,
  population.size.post.selection
)
}
\arguments{
\item{population.size.exposed.survivors}{= The relative population size of the mosquitoes in the intervention site who encountered the insecticide and survived.}

\item{mean.score.exposed.survivors}{= The mean Polygenic Resistance Score of the mosquitoes who encounter the insecticide and survive.}

\item{population.size.unexposed}{= The relative population size of the mosquitoes in the intervention site who did not encounter the insecticide.}

\item{trait.mean}{= The mean value of a polygenic trait before selection.}

\item{population.size.post.selection}{= The relative population size of all the mosquitoes in the intervention site after selection has occurred.}
}
\description{
Calculate the mean Polygenic Resistance Score of the all the mosquitoes after insecticide
selection has occurred.
}
