% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/can_insecticide_be_returned_to_arsenal.R
\name{can_insecticide_be_returned_to_arsenal}
\alias{can_insecticide_be_returned_to_arsenal}
\title{Check whether an insecticide should be returned back to the arsenal having been withdrawn.}
\usage{
can_insecticide_be_returned_to_arsenal(
  insecticide,
  current.generation,
  return.threshold,
  simulation.array
)
}
\arguments{
\item{insecticide}{= The insecticide to which the resistance intensity is to be checked against}

\item{current.generation}{= The generation where the simulation is up to.}

\item{return.threshold}{= The resistance intensity at which an insecticide should no longer be withdrawn; and returned to the arsenal.}

\item{simulation.array}{= The array which holds the simulation results.}
}
\value{
A TRUE or FALSE result.
}
\description{
Checks if a specified insecticide at a specified generation should be returned back to the arsenal for use.
This is based on whether the resistance intensity to this insecticide is less than or equal to the return threshold.
}
