% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bioassay_survival_to_field_survival.R
\name{convert_bioassay_survival_to_field_survival}
\alias{convert_bioassay_survival_to_field_survival}
\title{Convert bioassay survival to the expected field survival}
\usage{
convert_bioassay_survival_to_field_survival(
  bioassay.survival,
  regression.coefficient,
  regression.intercept,
  current.insecticide.efficacy
)
}
\arguments{
\item{bioassay.survival}{= The probability of a mosquito with a Polygenic Resistance Unit Score of z surviving in a WHO cylinder bioassay to the corresponding insecticide.}

\item{regression.coefficient}{= A linear model coefficient obtained from performing a linear model on paired experimental hut trials and WHO cylinder bioassays.}

\item{regression.intercept}{= The linear model intercept obtained from performing a linear model on paired experimental hut trials and WHO cylinder bioassays.}

\item{current.insecticide.efficacy}{= The insecticide efficacy of the insecticide at time since deployment defined as proportion of fully susceptible mosquitoes surviving contact with the insecticide-treated surface.}
}
\description{
Convert bioassay survival to the expected field survival
}
