% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_resistance_score_to_bioassay_survival.R
\name{convert_resistance_score_to_bioassay_survival}
\alias{convert_resistance_score_to_bioassay_survival}
\title{Convert Polygenic Resistance Score to bioassay survival proportion}
\usage{
convert_resistance_score_to_bioassay_survival(
  maximum.bioassay.survival.proportion = 1,
  trait.mean = 900,
  michaelis.menten.slope = 1,
  half.population.bioassay.survival.resistance = 900
)
}
\arguments{
\item{maximum.bioassay.survival.proportion}{Should be set as 1.}

\item{trait.mean}{The mean resistance intensity of the population.}

\item{michaelis.menten.slope}{Should be set as 1}

\item{half.population.bioassay.survival.resistance}{This is calculated using the calculate_half_population_resistance function}
}
\value{
bioassay.survival This is the mean bioassay survival for the population resistance intensity.
}
\description{
Convert Polygenic Resistance Score to bioassay survival proportion
}
