% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_normal_distribution.R
\name{create_normal_distribution}
\alias{create_normal_distribution}
\title{Generate a near perfect Normal Distribution}
\usage{
create_normal_distribution(vector.length, trait.mean, standard.deviation)
}
\arguments{
\item{vector.length}{= The length of the vector to be returned. A minimum value of 100,000 is recommmended.}

\item{trait.mean}{= The mean value of a polygenic trait in a population.}

\item{standard.deviaton}{= The standard deviation of the trait mean in the population.}
}
\description{
Create a reproducible and deterministic Normal distribution of trait values in a population.
}
