% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_male_total_population_size.R
\name{get_male_total_population_size}
\alias{get_male_total_population_size}
\title{Get the  total relative population size of male mosquitoes in the intervention site before insecticide selection.}
\usage{
get_male_total_population_size(total.population.size)
}
\arguments{
\item{total.population.size}{= The relative population size of the intervention site prior to selection.}
}
\description{
Get the  total relative population size of male mosquitoes in the intervention site before insecticide selection.
}
