% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_population_size_exposed_survivors.R
\name{get_population_size_exposed_survivors}
\alias{get_population_size_exposed_survivors}
\title{Calculate the relative population size of the mosquitoes in the intervention site who encountered the insecticide and survived}
\usage{
get_population_size_exposed_survivors(relative.contributions.after.selection)
}
\arguments{
\item{relative.contributions.after.selection}{= A vector containing the values of the relative population contributions to the population of mosquitoes for all Polygenic Resistance Scores in population after the insecticide encounter in the intervention site.}
}
\description{
Calculate the relative population size of the mosquitoes in the intervention site who encountered the insecticide and survived
}
