% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_population_size_unexposed.R
\name{get_population_size_unexposed}
\alias{get_population_size_unexposed}
\title{Calculate the relative population size of the mosquitoes in the intervention site who did not encounter the insecticide}
\usage{
get_population_size_unexposed(relative.contributions.unexposed)
}
\arguments{
\item{relative.contributions.unexposed}{= A vector containing the relative population contributions to the population of mosquitoes with a Polygenic Resistance Scores that did not encounter the insecticide in the intervention site.}
}
\description{
Calculate the relative population size of the mosquitoes in the intervention site who did not encounter the insecticide
}
