% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_total_population_size.R
\name{get_total_population_size}
\alias{get_total_population_size}
\title{Calculate the relative population size of the intervention site prior to selection.}
\usage{
get_total_population_size(relative.contributions.before.selection)
}
\arguments{
\item{relative.contributions.before.selection}{= a vector of probability densities calculated from calculate_density_of_trait_values}
}
\description{
Calculate the relative population size of the intervention site prior to selection.
}
