% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bioassay_survival_to_resistance.R
\name{\%>\%}
\alias{\%>\%}
\alias{plot_bioassay_survival_to_resistance}
\title{Plot the relationship between survival and resistance calculated from bioassay survival}
\usage{
plot_bioassay_survival_to_resistance(
  maximum.bioassay.survival.proportion = 1,
  michaelis.menten.slope = 1,
  half.population.bioassay.survival.resistance = 900,
  estimate.precision = 0.01,
  minimum.resistance.score = 0,
  maximum.resistance.score = 25000,
  minimum.bioassay.survival = 0,
  maximum.bioassay.survival = 1,
  divisions = 0.01
)
}
\arguments{
\item{maximum.bioassay.survival.proportion}{Should be set as 1.}

\item{michaelis.menten.slope}{Should be set as 1}

\item{half.population.bioassay.survival.resistance}{This is calculated using the calculate_half_population_resistance function}

\item{estimate.precision}{How precise your estimate of insecticide restistance intensity should be. Recommend values between 0.01 to 0.001}

\item{minimum.resistance.score}{Recommend setting as 0.}

\item{maximum.resistance.score}{This will depend on the half survival scale, but 25000 would be a good start.}

\item{minimum.bioassay.survival}{As a proportion, should be set as 0}

\item{maximum.bioassay.survival}{As a proportion, should be set as 1}

\item{divisions}{The resolution of the final plotted graph}
}
\value{
A ggplot of the relationship between resistance (x axis) and survival (y axis).
}
\description{
Plot the relationship between survival and resistance calculated from bioassay survival
}
