% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_migrating_intervention_to_refugia.R
\name{number_migrating_intervention_to_refugia}
\alias{number_migrating_intervention_to_refugia}
\title{Calculate the relative number of mosquitoes dispersing from the intervention site to the refugia.}
\usage{
number_migrating_intervention_to_refugia(dispersal.rate, intervention.coverage)
}
\arguments{
\item{dispersal.rate}{= The rate of mosquito exchange between the intervention site and the refugia.}

\item{intervention.coverage}{= The proportion of the total mosquito population that is covered by the intervention site.}
}
\description{
Calculate the relative number of mosquitoes dispersing from the intervention site to the refugia.
}
