% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_migrating_refugia_to_intervention.R
\name{number_migrating_refugia_to_intervention}
\alias{number_migrating_refugia_to_intervention}
\title{Calculate the relative number of mosquitoes dispersing from the refugia to the intervention site.}
\usage{
number_migrating_refugia_to_intervention(dispersal.rate, intervention.coverage)
}
\arguments{
\item{dispersal.rate}{= The rate of mosquito exchange between the intervention site and the refugia.}

\item{intervention.coverage}{= The proportion of the total mosquito population that is covered by the intervention site.}
}
\description{
Calculate the relative number of mosquitoes dispersing from the refugia to the intervention site.
}
