% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_resistance_to_bioassay_survival.R
\name{plot_resistance_to_bioassay_survival}
\alias{plot_resistance_to_bioassay_survival}
\title{Plot the relationship between resistance and survival calculated from insecticide resistance}
\usage{
plot_resistance_to_bioassay_survival(
  maximum.bioassay.survival.proportion = 1,
  michaelis.menten.slope = 1,
  half.population.bioassay.survival.resistance = 900,
  minimum.resistance = 0,
  maximum.resistance = 25000,
  resistance.score.divisions = 1
)
}
\arguments{
\item{maximum.bioassay.survival.proportion}{Should be set as 1.}

\item{michaelis.menten.slope}{Should be set as 1}

\item{half.population.bioassay.survival.resistance}{This is calculated using the calculate_half_population_resistance function}

\item{minimum.resistance}{Recommend setting as 0.}

\item{maximum.resistance}{This will depend on the half survival scale, but 10000 would be a good start.}

\item{bioassay.survival}{The survival that occured in the bioassay, as a proportion(values must be between 0 and 1). Where 1=all survived, 0=all died}

\item{estimate.precision}{How precise your estimate of insecticide restistance intensity should be. Recommend values between 0.01 to 0.001}

\item{sd.population.resistance}{How much variation in the population resistance.}

\item{nsim}{How many replications of the rnorm function are conducted. Recommended value is 1000.}
}
\value{
A ggplot of the relationship between resistance (x axis) and survival (y axis).
}
\description{
Plot the relationship between resistance and survival calculated from insecticide resistance
}
