% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_breeders_equation_male_female_fitness.R
\name{wrapper_breeders_equation_male_female_fitness}
\alias{wrapper_breeders_equation_male_female_fitness}
\title{Wrapper function to calculate the response when there is only fitness costs}
\usage{
wrapper_breeders_equation_male_female_fitness(
  heritability,
  trait.mean,
  female.fitness.cost,
  male.fitness.cost
)
}
\arguments{
\item{heritability}{= The heritability of a polygenic trait.}

\item{trait.mean}{= The mean value of a polygenic trait in a population.}

\item{female.fitness.cost}{= The fixed fitness cost associated with polygenic resistance for females.}

\item{male.fitness.cost}{= The fixed fitness costs associated with polygenic resistance for males}
}
\description{
Wrapper function to calculate the response when there is only fitness costs
}
