% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/wrapper_breeders_equation_male_female_insecticide_fitness.R
\name{wrapper_breeders_equation_male_female_insecticide_fitness}
\alias{wrapper_breeders_equation_male_female_insecticide_fitness}
\title{}
\usage{
wrapper_breeders_equation_male_female_insecticide_fitness(
  trait.mean,
  female.fitness.cost,
  male.fitness.cost,
  female.insecticide.exposure,
  male.insecticide.exposure,
  standard.deviation,
  vector.length,
  maximum.bioassay.survival.proportion,
  michaelis.menten.slope,
  half.population.bioassay.survival.resistance,
  regression.coefficient,
  regression.intercept,
  current.insecticide.efficacy,
  exposure.scaling.factor,
  heritability
)
}
\arguments{
\item{heritability}{= The heritability of a polygenic trait.}
}
\description{

}
