% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/wrapper_calculate_female_insecticide_fitness_selection_differential.R
\name{wrapper_calculate_female_insecticide_fitness_selection_differential}
\alias{wrapper_calculate_female_insecticide_fitness_selection_differential}
\title{A wrapper function to calculate the female insecticide and fitness cost selection differential}
\usage{
wrapper_calculate_female_insecticide_fitness_selection_differential(
  female.trait.mean,
  female.insecticide.exposure,
  standard.deviation,
  vector.length,
  maximum.bioassay.survival.proportion,
  michaelis.menten.slope,
  half.population.bioassay.survival.resistance,
  regression.coefficient,
  regression.intercept,
  current.insecticide.efficacy,
  exposure.scaling.factor
)
}
\arguments{
\item{female.trait.mean}{=}

\item{female.insecticide.exposure}{=}

\item{standard.deviation}{=}

\item{vector.length}{=}

\item{maximum.bioassay.survival.proportion}{=}

\item{michaelis.menten.slope}{=}

\item{half.population.bioassay.survival.resistance}{=}

\item{regression.coefficient}{=}

\item{regression.intercept}{=}

\item{current.insecticide.efficacy}{=}

\item{exposure.scaling.factor}{=}
}
\description{
A wrapper function to calculate the female insecticide and fitness cost selection differential
}
