% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/wrapper_calculate_male_insecticide_fitness_selection_differential.R
\name{wrapper_calculate_male_insecticide_fitness_selection_differential}
\alias{wrapper_calculate_male_insecticide_fitness_selection_differential}
\title{A wrapper function to calculate the male insecticide and fitness cost selection differential}
\usage{
wrapper_calculate_male_insecticide_fitness_selection_differential(
  male.trait.mean,
  female.insecticide.exposure,
  male.insecticide.exposure,
  standard.deviation,
  male.fitness.cost,
  vector.length,
  maximum.bioassay.survival.proportion,
  michaelis.menten.slope,
  half.population.bioassay.survival.resistance,
  regression.coefficient,
  regression.intercept,
  current.insecticide.efficacy,
  exposure.scaling.factor
)
}
\arguments{
\item{male.trait.mean}{= The mean Polygenic Resistance Score to the insecticide for male mosquitoes before selection has occurred.}

\item{female.insecticide.exposure}{=Proportion of female mosquitoes in the intervention site that encounter and are exposed to the deployed insecticide.}

\item{male.insecticide.exposure}{= Proportion of male mosquitoes in the intervention site that encounter and are exposed to the deployed insecticide as a proportion of the exposure of female mosquitoes}

\item{standard.deviation}{= The standard deviation of the trait mean in the population.}

\item{vector.length}{= The length of the vector to be returned. A minimum value of 100,000 is recommmended.}

\item{maximum.bioassay.survival.proportion}{= The maximum proportion of mosquitoes that can survive in the bioassay.}

\item{michaelis.menten.slope}{= The slope in the Michaelis-Menten equation.}

\item{half.population.bioassay.survival.resistance}{= The Polygenic Resistance Score which gives a 50\% survival probability in a WHO cylinder bioassay.}

\item{regression.coefficient}{= A linear model coefficient obtained from performing a linear model on paired experimental hut trials and WHO cylinder bioassays.}

\item{regression.intercept}{= The linear model intercept obtained from performing a linear model on paired experimental hut trials and WHO cylinder bioassays.}

\item{current.insecticide.efficacy}{= The current efficacy of the insecticide, defined as proportion of fully susceptible mosquitoes surviving contact with the insecticide-treated surface.}

\item{exposure.scaling.factor}{= A factor which converts the insecticide exposure to the selection differential.}
}
\description{
A wrapper function to calculate the male insecticide and fitness cost selection differential
}
