% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_female_fitness_selection_differential.R
\name{wrapper_female_fitness_selection_differential}
\alias{wrapper_female_fitness_selection_differential}
\title{A wrapper function to calculate the fitness selection differential for female mosquitoes}
\usage{
wrapper_female_fitness_selection_differential(
  female.trait.mean,
  female.fitness.cost
)
}
\arguments{
\item{female.trait.mean}{= The mean Polygenic Resistance Score of female mosquitoes before selection has occurred.}

\item{female.fitness.cost}{= The fixed fitness costs associated with polygenic resistance for females}
}
\description{
A wrapper function to calculate the fitness selection differential for female mosquitoes
}
