% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/wrapper_intervention_site_after_selection_deployed.R
\name{wrapper_intervention_site_after_selection_deployed}
\alias{wrapper_intervention_site_after_selection_deployed}
\title{Wrapper function to get the Mean Polygenic Resistance Score  in the Intervention Site after Selection}
\usage{
wrapper_intervention_site_after_selection_deployed(
  intervention.before.selection,
  female.fitness.cost,
  male.fitness.cost,
  female.insecticide.exposure,
  male.insecticide.exposure,
  standard.deviation,
  vector.length,
  maximum.bioassay.survival.proportion,
  michaelis.menten.slope,
  half.population.bioassay.survival.resistance,
  regression.coefficient,
  regression.intercept,
  current.insecticide.efficacy,
  exposure.scaling.factor,
  heritability
)
}
\arguments{
\item{intervention.before.selection}{= The mean Polygenic Resistance Score of the mosquito population in the intervention before selection has occurred that generation.}

\item{female.fitness.cost}{= The fixed fitness cost associated with polygenic resistance for females.}

\item{male.fitness.cost}{= The fixed fitness cost associated with polygenic resistance for males.}

\item{female.insecticide.exposure}{= Proportion of female mosquitoes in the intervention site that encounter and are exposed to the deployed insecticide.}

\item{male.insecticide.exposure}{= Proportion of male mosquitoes in the intervention site that encounter and are exposed to the deployed insecticide as a proportion of the exposure of female mosquitoes}

\item{standard.deviation}{= The standard deviation of the trait mean in the population.}

\item{vector.length}{= The length of the vector to be returned. A minimum value of 100,000 is recommmended.}

\item{maximum.bioassay.survival.proportion}{= The maximum proportion of mosquitoes that can survive in the bioassay.}

\item{michaelis.menten.slope}{= The slope in the Michaelis-Menten equation}

\item{half.population.bioassay.survival.resistance}{= The Polygenic Resistance Score which gives a 50\% survival probability in a WHO cylinder bioassay.}

\item{regression.coefficient}{= A linear model coefficient obtained from performing a linear model on paired experimental hut trials and WHO cylinder bioassays.}

\item{regression.intercept}{= The linear model intercept obtained from performing a linear model on paired experimental hut trials and WHO cylinder bioassays.}

\item{current.insecticide.efficacy}{= The insecticide efficacy of insecticide i at time since deployment τ defined as proportion of fully susceptible mosquitoes surviving contact with the insecticide-treated surface.,}

\item{exposure.scaling.factor}{= A factor which converts the insecticide exposure to the selection differential.}

\item{heritability}{= The heritability of a polygenic trait.}
}
\description{
Wrapper function to get the Mean Polygenic Resistance Score  in the Intervention Site after Selection
}
