% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/wrapper_intervention_site_after_selection_not_deployed.R
\name{wrapper_intervention_site_after_selection_not_deployed}
\alias{wrapper_intervention_site_after_selection_not_deployed}
\title{A wrapper function to calculate the trait mean in the intervention site after fitness cost based selection}
\usage{
wrapper_intervention_site_after_selection_not_deployed(
  heritability,
  intervention.before.selection,
  female.fitness.cost,
  male.fitness.cost
)
}
\arguments{
\item{intervention.before.selection}{= The mean Polygenic Resistance Score of the mosquito population in the intervention before selection has occurred that generation.}

\item{female.fitness.cost}{= The fixed fitness cost associated with polygenic resistance for females.}

\item{male.fitness.cost}{= The fixed fitness costs associated with polygenic resistance for males}

\item{heritabililty}{= The heritability of a polygenic trait.}
}
\description{
A wrapper function to calculate the trait mean in the intervention site after fitness cost based selection
}
