% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_intervention_site_deployed_dispersal.R
\name{wrapper_intervention_site_deployed_dispersal}
\alias{wrapper_intervention_site_deployed_dispersal}
\title{Wrapper to allow mosquito dispersal tracking in the intervention site when the insecticide is deployed}
\usage{
wrapper_intervention_site_deployed_dispersal(
  insecticide.population.suppression,
  intervention.before.selection,
  female.fitness.cost,
  male.fitness.cost,
  female.insecticide.exposure,
  male.insecticide.exposure,
  standard.deviation,
  vector.length,
  maximum.bioassay.survival.proportion,
  michaelis.menten.slope,
  half.population.bioassay.survival.resistance,
  regression.coefficient,
  regression.intercept,
  current.insecticide.efficacy,
  exposure.scaling.factor,
  heritability,
  refugia.before.selection,
  dispersal.rate,
  intervention.coverage
)
}
\arguments{
\item{insecticide.population.suppression}{= The impact of insecticides on the relative population size of female mosquitoes the intervention site, who are now ready to lay eggs.}

\item{intervention.before.selection}{= The mean Polygenic Resistance Score of the mosquito population in the intervention before selection has occurred that generation.}

\item{female.fitness.cost}{= The fixed fitness cost associated with polygenic resistance for females.}

\item{male.fitness.cost}{= The fixed fitness cost associated with polygenic resistance for males.}

\item{female.insecticide.exposure}{= Proportion of female mosquitoes in the intervention site that encounter and are exposed to the deployed insecticide.}

\item{male.insecticide.exposure}{= Proportion of male mosquitoes in the intervention site that encounter and are exposed to the deployed insecticide as a proportion of the exposure of female mosquitoes}

\item{standard.deviation}{= The standard deviation of the trait mean in the population.}

\item{vector.length}{= The length of the vector to be returned. A minimum value of 100,000 is recommmended.}

\item{maximum.bioassay.survival.proportion}{= The maximum proportion of mosquitoes that can survive in the bioassay.}

\item{michaelis.menten.slope}{= The slope in the Michaelis-Menten equation}

\item{half.population.bioassay.survival.resistance}{= The Polygenic Resistance Score which gives a 50\% survival probability in a WHO cylinder bioassay.}

\item{regression.coefficient}{= A linear model coefficient obtained from performing a linear model on paired experimental hut trials and WHO cylinder bioassays.}

\item{regression.intercept}{= The linear model intercept obtained from performing a linear model on paired experimental hut trials and WHO cylinder bioassays.}

\item{current.insecticide.efficacy}{= The insecticide efficacy of insecticide i at time since deployment τ defined as proportion of fully susceptible mosquitoes surviving contact with the insecticide-treated surface.,}

\item{exposure.scaling.factor}{= A factor which converts the insecticide exposure to the selection differential.}

\item{heritability}{= The heritability of a polygenic trait.}

\item{refugia.before.selection}{= The mean Polygenic Resistance Score of the mosquito population in the refugia  before selection has occurred that generation.}

\item{dispersal.rate}{= The rate of mosquito exchange between the intervention site and the refugia.}

\item{intervention.coverage}{= The proportion of the total mosquito population that is covered by the intervention site.}
}
\description{
Wrapper to allow mosquito dispersal tracking in the intervention site when the insecticide is deployed
}
