% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/wrapper_intervention_site_not_deployed_dispersal.R
\name{wrapper_intervention_site_not_deployed_dispersal}
\alias{wrapper_intervention_site_not_deployed_dispersal}
\title{Wrapper to allow mosquito dispersal tracking in the intervention site when the insecticide is not deployed}
\usage{
wrapper_intervention_site_not_deployed_dispersal(
  insecticide.population.suppression,
  intervention.before.selection,
  female.fitness.cost,
  male.fitness.cost,
  heritability,
  refugia.before.selection,
  dispersal.rate,
  intervention.coverage
)
}
\arguments{
\item{intervention.before.selection}{= The mean Polygenic Resistance Score of the mosquito population in the intervention before selection has occurred that generation.}

\item{female.fitness.cost}{= The fixed fitness cost associated with polygenic resistance for females.}

\item{male.fitness.cost}{= The fixed fitness costs associated with polygenic resistance for males}

\item{refugia.before.selection}{= The mean Polygenic Resistance Score of the mosquito population in the refugia  before selection has occurred that generation.}

\item{dispersal.rate}{= The rate of mosquito exchange between the intervention site and the refugia.}

\item{intervention.coverage}{= The proportion of the total mosquito population that is covered by the intervention site.}

\item{heritabililty}{= The heritability of a polygenic trait.}
}
\description{
Wrapper to allow mosquito dispersal tracking in the intervention site when the insecticide is not deployed
}
