% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_male_fitness_selection_differential.R
\name{wrapper_male_fitness_selection_differential}
\alias{wrapper_male_fitness_selection_differential}
\title{A wrapper function to calculate the fitness selection differential for male mosquitoes}
\usage{
wrapper_male_fitness_selection_differential(male.trait.mean, male.fitness.cost)
}
\arguments{
\item{male.trait.mean}{= The mean Polygenic Resistance Score of male mosquitoes before selection has occurred.}

\item{male.fitness.cost}{= The fixed fitness costs associated with polygenic resistance for males}
}
\description{
A wrapper function to calculate the fitness selection differential for male mosquitoes
}
