% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/wrapper_male_insecticide_selection_differential.R
\name{wrapper_male_insecticide_selection_differential}
\alias{wrapper_male_insecticide_selection_differential}
\title{A wrapper function to calculate the insecticide selection differential for male mosquitoes}
\usage{
wrapper_male_insecticide_selection_differential(
  intervention.before.selection,
  male.insecticide.exposure,
  female.insecticide.exposure,
  standard.deviation,
  vector.length,
  maximum.bioassay.survival.proportion = 1,
  michaelis.menten.slope = 1,
  half.population.bioassay.survival.resistance = 900,
  regression.coefficient = 0.48,
  regression.intercept = 0.15,
  current.insecticide.efficacy
)
}
\arguments{
\item{intervention.before.selection}{= The mean Polygenic Resistance Score of the male mosquitoes prior to insecticide selection this generation.}

\item{male.insecticide.exposure}{= Proportion of male mosquitoes in the intervention site that encounter and are exposed to the deployed insecticide as a proportion of the exposure of female mosquitoes}

\item{female.insecticide.exposure}{= Proportion of female mosquitoes in the intervention site that encounter and are exposed to the deployed insecticide.}

\item{standard.deviation}{= The standard deviation of the trait mean in the population.}

\item{vector.length}{= The length of the vector to be returned. A minimum value of 100,000 is recommmended.}

\item{maximum.bioassay.survival.proportion}{= The maximum proportion of mosquitoes that can survive in the bioassay.}

\item{michaelis.menten.slope}{= The slope in the Michaelis-Menten equation.}

\item{half.population.bioassay.survival.resistance}{= The Polygenic Resistance Score which gives a 50\% survival probability in a WHO cylinder bioassay.}

\item{regression.coefficient}{= A linear model coefficient obtained from performing a linear model on paired experimental hut trials and WHO cylinder bioassays.}

\item{regression.intercept}{= The linear model intercept obtained from performing a linear model on paired experimental hut trials and WHO cylinder bioassays.}

\item{current.insecticide.efficacy}{= The current efficacy of the insecticide, defined as proportion of fully susceptible mosquitoes surviving contact with the insecticide-treated surface.}

\item{trait.mean}{= The mean value of a polygenic trait in a population.}
}
\description{
A wrapper function to calculate the insecticide selection differential for male mosquitoes
}
