% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_refugia_breeders_equation.R
\name{wrapper_refugia_breeders_equation}
\alias{wrapper_refugia_breeders_equation}
\title{Wrapper to perfrom the Breeders equation in the Refugia}
\usage{
wrapper_refugia_breeders_equation(
  refugia.before.selection,
  heritability,
  female.fitness.cost,
  male.fitness.cost
)
}
\arguments{
\item{refugia.before.selection}{= The mean Polygenic Resistance Score of the mosquito population in the refugia  before selection has occurred that generation.}

\item{heritability}{= The heritability of a polygenic trait.}

\item{female.fitness.cost}{= The fixed fitness cost associated with polygenic resistance for females.}

\item{male.fitness.cost}{= The fixed fitness costs associated with polygenic resistance for males'}
}
\description{
Wrapper to perfrom the Breeders equation in the Refugia
}
