% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_total_population_size.R
\name{wrapper_total_population_size}
\alias{wrapper_total_population_size}
\title{A wrapper function for calculating the relative total population size}
\usage{
wrapper_total_population_size(standard.deviation, vector.length)
}
\arguments{
\item{vector.length}{= The length of the vector to be returned. A minimum value of 100,000 is recommmended.}

\item{standard.deviaton}{= The standard deviation of the trait mean in the population.}
}
\description{
A wrapper function for calculating the relative total population size
}
